+++ sdk/lib/developer/timeline.dart
@@ -190,12 +190,8 @@ class TimelineTask {
 
   /// Create a task with an explicit [taskId]. This is useful if you are
   /// passing a task from one isolate to another.
-  ///
-  /// If [parent] is provided, the parent's task ID is provided as argument
-  /// 'parentId' when [start] is called. In DevTools, this argument will result
-  /// in this [TimelineTask] being linked to the [parent] [TimelineTask].
-  TimelineTask.withTaskId(int taskId, {TimelineTask parent})
-      : _parent = parent,
+  TimelineTask.withTaskId(int taskId)
+      : _parent = null,
         _taskId = taskId {
     ArgumentError.checkNotNull(taskId, 'taskId');
   }